Integrate the cropping with the Telegram bot (#28)
* Minor changes.

* Make the bot async.

* remove duplicated requirement

* fix requirementes.txt

* Fix Typos.

* Minor changes.

* Add asyncio to the text_to_audio function

* Integrate everything so far into the bot.

* Some refactoring.

* Minor changes.

* Integrate all models in the bot. Do some refactor.

* Minor change / refactor.

* Minor change.

* Minor fix on data/.gitignore

* Normalize the vectors on the Index Tree.

* Add cropping layer to the process of uploading a picture.

* Add cropping layer to the process of uploading a picture.

* Minor change. Removed line that was printing the image every time it was processed.

+++ reverse_image_search_module/resnet18.py
@@ -24,7 +24,7 @@ class Img2VecResnet18():
         return cnnModel, layer
 
     def getVec(self, img):
-        image = self.normalize(self.toTensor(img)).unsqueeze(0).to(self.device)
+        image = self.normalize(self.toTensor(img)).unsqueeze(0).to(torch.float32).to(self.device)
         embedding = torch.zeros(1, self.numberFeatures, 1, 1)
 
         def copyData(m, i, o): embedding.copy_(o.data)
@@ -37,5 +37,6 @@ class Img2VecResnet18():
     def getNormalizedVec(self, img):
         vec = self.getVec(img)
         return vec / np.linalg.norm(vec)
+
 
 img2vec = Img2VecResnet18()
